refactor: Consolidate duplicate test fake repository implementations
- Create shared test utilities package at util/fakes/
- Implement consolidated FakeScoreRepository with lastSavedState tracking
- Implement consolidated FakeSettingsRepository with complete helper methods
- Implement consolidated FakeMatchRepository with append-only behavior
- Update all test files to use shared implementations
- Remove ~152 lines of duplicate code across 5 test files

This creates a single source of truth for test fakes, making them easier
to maintain and ensuring consistent behavior across all tests.

Closes #3
