+++ app/src/test/java/com/soyvictorherrera/scorecount/ui/matchhistory/MatchHistoryViewModelTest.kt
@@ -3,10 +3,10 @@ package com.soyvictorherrera.scorecount.ui.matchhistory
 import com.soyvictorherrera.scorecount.domain.model.Match
 import com.soyvictorherrera.scorecount.domain.repository.MatchRepository
 import com.soyvictorherrera.scorecount.domain.usecase.GetMatchesUseCase
+import com.soyvictorherrera.scorecount.util.fakes.FakeMatchRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.resetMain
@@ -181,17 +181,5 @@ class MatchHistoryViewModelTest {
         assertEquals("1", matches[0].id)
         assertEquals("2", matches[1].id)
         assertEquals("3", matches[2].id)
-    }
-
-    // --- Fake Repository ---
-
-    class FakeMatchRepository : MatchRepository {
-        private val _matches = MutableStateFlow<List<Match>>(emptyList())
-
-        override fun getMatchList(): Flow<List<Match>> = _matches
-
-        override suspend fun saveMatch(match: Match) {
-            _matches.value = _matches.value + match
-        }
     }
 }
